requirements.txt
Pillow==3.2.0
pytest==2.9.1
